Add TriHex solver.
This solver generates a uniform-ish lattice of colors with the power of Hexagons(tm).
The colors it generates are located in three tiers of brightness. Within each tier
colors share a colorfulness and are equally spaces in hue. This lattice has the property
that all nearest-neighbor distances between rings are equal, and this is the same distance
as nearest-neighbor distance within the lowest ring.